Active passes, reset pass, extend pass
!!!!!!! Guys please check the database 5 times while review, I am not sure if everything about migrations is ok
added methods to return current active passes, to deactivate a chosen pass or to extend time of a chosen pass
