+++ GymApp/GymApp/Controllers/ClientsController.cs
@@ -21,54 +21,6 @@ namespace GymApp.Controllers
             _context = context;
         }
 
-        // get all the active passes
-        [HttpGet("GetActivePasses/{clientId}")]
-        public IActionResult GetActivePasses(Guid clientId)
-        {
-            var activePassBoughtEvents = _context.Client_1
-                .Where(c => c.Id == clientId)
-                .SelectMany(c => c.PassBoughtEvents)
-                .Include(pbe => pbe.Pass)
-                .Where(pbe => DateTime.UtcNow < pbe.DateTime.AddDays(pbe.Pass.Duration * pbe.refresh))
-                .ToList();
-
-            return Ok(activePassBoughtEvents);
-        }
-
-
-        //reseting the Pass
-        [HttpPut("ResetPassBoughtEventRefresh/{passBoughtEventId}")]
-        public IActionResult ResetPassBoughtEventRefresh(Guid passBoughtEventId)
-        {
-            var passBoughtEvent = _context.PassBoughtEvent_1.FirstOrDefault(pbe => pbe.Id == passBoughtEventId);
-            if (passBoughtEvent == null)
-            {
-                return NotFound();
-            }
-
-            passBoughtEvent.refresh = 0;
-            _context.SaveChanges();
-
-            return Ok();
-        }
-
-        // Extend the pass validity
-        [HttpPut("ExtendPassValidity/{passBoughtEventId}")]
-        public IActionResult ExtendPassValidity(Guid passBoughtEventId, int extension)
-        {
-            var passBoughtEvent = _context.PassBoughtEvent_1.FirstOrDefault(pbe => pbe.Id == passBoughtEventId);
-            if (passBoughtEvent == null)
-            {
-                return NotFound();
-            }
-
-            passBoughtEvent.Pass.Duration += extension;
-            _context.SaveChanges();
-
-            return Ok();
-        }
-
-
         // GET: api/Clients
         [HttpGet()]
         public async Task<ActionResult<IEnumerable<Client>>> GetClient_1()
